% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{browser}
\alias{browser}
\title{Browser}
\usage{
browser(
  dataset = NULL,
  categoryFilter = "Presidential Actions",
  contentType = "AI_Summary",
  row = NULL,
  mode = "first"
)
}
\arguments{
\item{dataset}{The WH data object}

\item{categoryFilter}{A category id to filter by.}

\item{contentType}{Content type filter.}

\item{row}{Select a particular folder.}

\item{mode}{Either 'first' or 'last'}
}
\value{
An interactive text object.
}
\description{
An interactive data browser to navigate WH scraped files.
}
\examples{
\dontrun{
browser(data)
}
}
