% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{charclean}
\alias{charclean}
\title{charclean}
\usage{
charclean(vec)
}
\arguments{
\item{vec}{String, a character value to be cleaned and reformatted.}
}
\value{
A strong.
}
\description{
A utility function to create a character value that strips away possible formatting changes to match strings, such as vendor strings.
}
\examples{
\dontrun{
charclean("Vendor Name. Inc")
}

}
