% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{check_bracket_mismatch}
\alias{check_bracket_mismatch}
\title{check_bracket_mismatch
A utility function to help with JSON bracket mismatches.}
\usage{
check_bracket_mismatch(json_string)
}
\arguments{
\item{json_string}{A preJSON formatted string.}
}
\value{
a dataframe
}
\description{
check_bracket_mismatch
A utility function to help with JSON bracket mismatches.
}
\examples{
\dontrun{
check_bracket_mismatch(scripts)
}
}
