% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{check_xml_type}
\alias{check_xml_type}
\title{check_xml_type}
\usage{
check_xml_type(doc, return = "check")
}
\arguments{
\item{doc}{An XML doc.}

\item{return}{Either "check" or "master", returning the type or a master list of types.}
}
\value{
A string, or list.
}
\description{
A utility function to set type and extract proper xml node for FPDS data.
}
\examples{
\dontrun{
check_xml_type(xdoc,return="check")
}

}
