% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{clean_and_match}
\alias{clean_and_match}
\title{clean_and_match}
\usage{
clean_and_match(
  FPDS_data = NULL,
  DOGE_data = NULL,
  returnUnmatched = FALSE,
  returnDuplicates = FALSE,
  print_status = TRUE,
  verbose = F
)
}
\arguments{
\item{FPDS_data}{Supplied data from FPDS}

\item{DOGE_data}{Supplied data from DOGE}

\item{returnUnmatched}{Logical, whether to return unmatched PIIDS}

\item{returnDuplicates}{Logical, whether to return duplciate PIIDS}

\item{print_status}{Logical, whether to return feedback}

\item{verbose}{Logical, whether to return messages}
}
\value{
A named list with data.frames.
}
\description{
A utility function designed to help match DOGE data and FPDS data. FPDS data
returns multiple transaction values per PIID; DOGE data only includes on row.
While it is possible to match on PIID and modNumber, this frequently returns
several values. Additionally, descriptions and vendors do not always match,
as there seems to be some editing that occassionally gets in the way. This
function seeks to standardize and match data using iterative matching via
increasingly broadening scope. Duplicates are dealt with by repeating the
process if possible.
}
\examples{
\dontrun{
clean_and_match(F,D,print_status=TRUE)
}
}
