% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{clone_dataframe_structure}
\alias{clone_dataframe_structure}
\title{clone_dataframe_structure}
\usage{
clone_dataframe_structure(source = NULL, clone = NULL)
}
\arguments{
\item{source}{Primary data.frame to source from.}

\item{clone}{Data.frame to clone.}
}
\value{
A data.frame.
}
\description{
Internal function to clone structure of a dataframe.
}
\examples{
\dontrun{
match_col_types(a,b)
}

}
