% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{convertability}
\alias{convertability}
\title{convertability}
\usage{
convertability(vec, type)
}
\arguments{
\item{vec}{A vector consisting of unchecked values.}

\item{type}{String, the type to check for}
}
\value{
String with value of type.
}
\description{
Simple utility function to determine whether a vector is all of the same type.
}
\examples{
\dontrun{
convertability(vector, type="character")
}

}
