% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{data_checker}
\alias{data_checker}
\title{data_checker}
\usage{
data_checker(data = NULL, sub = NULL, var = NULL, unique = FALSE)
}
\arguments{
\item{data}{Supplied data}

\item{sub}{A sub element of a list. Name.}

\item{var}{A variable of a data.frame.}

\item{unique}{Logical, to render unique values}
}
\value{
Either character vector or data.frame.
}
\description{
Utility function to check and return proper data efficiently
}
\examples{
\dontrun{
data_checker(F,"contracts","PIID")
}
}
