% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{doge_api_call}
\alias{doge_api_call}
\title{doge_api_call}
\usage{
doge_api_call(
  path = NULL,
  kind = NULL,
  limit = NULL,
  pnum = NULL,
  verbose = FALSE
)
}
\arguments{
\item{path}{A vector of paths to call.}

\item{kind}{Kind of data}

\item{limit}{Items per page; defaults to 500}

\item{pnum}{Page number to start at}
}
\value{
A dataframe
}
\description{
Internal utility function to capture DOGE api data.
}
\examples{
\dontrun{
doge_api_call()
}
}
