% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doge.R
\name{doge_get_data}
\alias{doge_get_data}
\title{DOGE Get Data}
\usage{
doge_get_data(verbose = TRUE)
}
\arguments{
\item{verbose}{}
}
\value{
A named list comprised of dataframes.
}
\description{
A fairly straightforward scraper that pulls data from dynamically-created
scripts (javascript), cleans, and processes in table form.
}
\examples{
get_doge_data()

}
