% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{doge_get_schema}
\alias{doge_get_schema}
\title{doge_get_schema}
\usage{
doge_get_schema()
}
\value{
An XML schema
}
\description{
Utility function to capture DOGE schema from api.doge.gov.
}
\examples{
\dontrun{
doge_get_schema()
}

}
