% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doge.R
\name{doge_summarize}
\alias{doge_summarize}
\title{DOGE Summarize}
\usage{
doge_summarize(data = NULL, return = "all")
}
\arguments{
\item{data}{}

\item{return}{}
}
\value{
A printed data.frame.
}
\description{
Present basic stats on the DOGE data scraped.
}
\examples{
\dontrun{
doge_summarize(data)
}
}
