% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{explode_urls}
\alias{explode_urls}
\title{explode_urls}
\usage{
explode_urls(y, k)
}
\arguments{
\item{y}{The url list (or a doge object data.frame)}

\item{k}{The kind of dataset (contracts, grants, leases, payments)}
}
\value{
a dataframe of exploded url components
}
\description{
Utility function to extract url parameters to dataframe.
}
\examples{
\dontrun{
explode_urls(urls,"contracts")
}
}
