% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{extract_outlays}
\alias{extract_outlays}
\title{extract_outlays}
\usage{
extract_outlays(x = NULL, type = NULL)
}
\arguments{
\item{x}{Dataframe containing outlay data.}

\item{type}{The type of award to analyze.}
}
\value{
A data.frame
}
\description{
A utility function to extract outlays from ungrouped data captured via usaspending API.
}
\examples{
\dontrun{
extract_outlays(outlay_data,type="AWARD")
}

}
