% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{extract_url_components}
\alias{extract_url_components}
\title{Extract URL Components}
\usage{
extract_url_components(data = NULL)
}
\arguments{
\item{data}{}

\item{url_col}{}

\item{link}{}
}
\value{
A data.frame
}
\description{
Internal function designed to pull out url components in doge fpds_link field.
}
\examples{
\dontrun{
extract_url_components(df)
}

}
