% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpds.R
\name{fpds_call}
\alias{fpds_call}
\title{fpds_call}
\usage{
fpds_call(piid_val = NULL, start = NULL, return = "xml", verbose = FALSE)
}
\arguments{
\item{piid_val}{Supplied PIID lookup}

\item{start}{For pagination, start value}

\item{return}{Options are "xml", "url," "raw." Url returns call url; xml fetches the xml needed for parsing; raw is for testing.}

\item{verbose}{Logical, whether to provide feedback.}
}
\value{
Either an xml file, a raw html content file, or the url.
}
\description{
fpds_call
}
\examples{
\dontrun{
fpds_call("XXXXXXX")
}
}
