% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{fpds_get_new}
\alias{fpds_get_new}
\title{FPDS Get New}
\usage{
fpds_get_new(
  doge_data = NULL,
  fpds_data = NULL,
  verbose = FALSE,
  return = "combined"
)
}
\arguments{
\item{doge_data}{A data object created by doge_get_data()}

\item{fpds_data}{A data object created by fpds_get_data()}

\item{verbose}{Logical, whether to return messages.}

\item{return}{Whether to return a merged file with new and old data, or
simply the new data.}
}
\value{
A named list comrpised of data.frames.
}
\description{
A convenience function to compare a recently captured DOGE object with data
previously downloaded by FPDS. This will run requests to FPDS based only on
new DOGE rows captured
}
\examples{
\dontrun{
fpds_get_new(doge_data,fpds_data)
}
}
