% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpds.R
\name{fpds_stats}
\alias{fpds_stats}
\title{FPDS Stats}
\usage{
fpds_stats(df, returnSummary = TRUE, returnUniquePIIDs = FALSE)
}
\arguments{
\item{df}{The data object returned by fpds_get_data().}

\item{returnSummary}{Logical, whether to return stat summary.}

\item{returnUniquePIIDs}{Logical, whether to return unique PIID values.}
}
\value{
A printed data.frame, with optional vector of PIID values.
}
\description{
A simple function to return quick and dirty stats on the FPDS data
downloaded. Particularly useful for determining if nrow() was achieved as
intended, and to compare to data captured by doge scraper.
}
\examples{
fpds_stats(t,returnSummary=T,returnUniquePIIDS=F)
}
