% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpds.R
\name{fpds_table}
\alias{fpds_table}
\title{FPDS Table}
\usage{
fpds_table(
  x,
  ns = NULL,
  ns_pre = "ns1",
  piid = NULL,
  idv_piid = NULL,
  contract_type = NULL
)
}
\arguments{
\item{x}{A valid XML object}

\item{ns}{Namespace generated by fpds_get_data()}

\item{ns_pre}{Namespace prefix}

\item{piid}{PIID values passed by get_fdps_data}

\item{idv_piid}{}

\item{contract_type}{Contract type (c("AWARD","IDV")) passed from fpds_get_data()}
}
\value{
A data.frame.
}
\description{
A function to read FDPS XML data, clean it, and create table data. Since
"award" and "IDV" queries return slightly different XML tags, we discriminate
between the two returns for tabulature.
}
\examples{
\dontrun{
fpds_table(x,ns,piid=piid_val,idv_piid=idv_PIID_val,contract_type=ct)
}

}
