% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_usaspending_outlay_loop}
\alias{get_usaspending_outlay_loop}
\title{get_usaspending_outlay_loop}
\usage{
get_usaspending_outlay_loop(merged_data = NULL, progress_style = 1)
}
\arguments{
\item{merged_data}{Data created by clean_and_match(). Cycles through unique_ids in data.}
}
\value{
A data.frame
}
\description{
A utility function to run get_usaspending_outlay on a loop of provided unique_ids.
}
\examples{
\dontrun{
get_usaspending_outlay_loop(M)
}

}
