% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{load_FPDS_fields}
\alias{load_FPDS_fields}
\title{Load FPDS Fields}
\usage{
load_FPDS_fields(contract_type = NULL)
}
\arguments{
\item{contract_type}{}
}
\value{
A list.
}
\description{
A value placeholder for grabbing desired XML fields.
}
