% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{match_col_types}
\alias{match_col_types}
\title{match_col_types}
\usage{
match_col_types(master = NULL, slave = NULL)
}
\arguments{
\item{master}{Primary data.frame to source from.}

\item{slave}{Data.frame to clone.}
}
\value{
A data.frame.
}
\description{
Internal function to clone structure of a dataframe.
}
\examples{
\dontrun{
match_col_types(a,b)
}

}
