% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oa.R
\name{oa_query}
\alias{oa_query}
\title{oa_query}
\usage{
oa_query(prompt = NULL, api_key = NULL, model = "gpt-4o-mini")
}
\arguments{
\item{prompt}{A text prompt sent via chat_openai().}

\item{api_key}{Api key supplied if not set as environmental variable.}

\item{model}{Select the proper mode. Recommended (and default): gpt-4o-mini for cost effectiveness.}
}
\value{
Text object
}
\description{
Wrapper for ellmer:chat_openai(), binding prompt, api key, and model.
}
\examples{
\dontrun{
oa_query("Tell me a story", api_key="XXXXXXX",model="gpt-4o-mini")
}
}
