% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oa.R
\name{oa_summary}
\alias{oa_summary}
\title{oa_summary()}
\usage{
oa_summary(
  text = NULL,
  summary_prompt = NULL,
  summary_sentence_length = 5,
  api_key = NULL,
  model = "gpt-4o-mini"
)
}
\arguments{
\item{text}{Text to be summarized. NOTE: Depending on the model, you are limited to your token size. gpt-4o-mini, for example is limited to 200,000 TPM.}

\item{summary_prompt}{A user-supplied summary prompt. Default is "Generate a brief summary of the following text."}

\item{summary_sentence_length}{A user-supplied limit on summary length. Default is 5 sentences.}

\item{api_key}{Api key supplied if not set as environmental variable.}

\item{model}{Select the proper mode. Recommended (and default): gpt-4o-mini for cost effectiveness.}
}
\value{
A text response.
}
\description{
A packaged summary request for oa_query. Can be modified with two user args.
}
\examples{
oa_summary(text="This is a bunch of text.")
}
