% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{pkg_merge}
\alias{pkg_merge}
\title{pkg_merge}
\usage{
pkg_merge(old_data = NULL, new_data = NULL)
}
\arguments{
\item{old_data}{Dataframe, previous FPDS data}

\item{new_data}{Dataframe, new FPDS data}
}
\value{
Dataframe
}
\description{
A utility function to accurately merge FPDS data.
}
\examples{
\dontrun{
pkg_merge(old_data, new_data)
}

}
