% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oa.R
\name{set_oa_api_key}
\alias{set_oa_api_key}
\title{set_oa_api_key()}
\usage{
set_oa_api_key(key = NULL)
}
\arguments{
\item{key}{An api key}
}
\value{
An environmental variable
}
\description{
Sets an environmental variable named OA_KEY with your OpenAI api key.
}
\examples{
set_oa_api_key("XXXXXXXXXXXXX")

}
