% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{usaspending_api_call}
\alias{usaspending_api_call}
\title{usaspending_api_call}
\usage{
usaspending_api_call(
  unique_id = NULL,
  type = NULL,
  formatted = T,
  limit = 50,
  page = 1,
  verbose = F,
  returnRaw = F
)
}
\arguments{
\item{unique_id}{A unique award key. Can be generated using generate_unique_key.}

\item{type}{The type of call. Currently support funding, funding_rollup, download, accounts, awards, award_summary, and amounts.}

\item{formatted}{Logical: whether to return a dataframe or leave as JSON list if so supplied.}

\item{limit}{Integer: how many items per page requested.}

\item{page}{Integer: if paginated, which page to call.}

\item{verbose}{Logical: Whether to provide feedback.}

\item{returnRaw}{Logical: Whether to provide raw call straight from GET/POST.}
}
\value{
A data.frame
}
\description{
R wrapper for api interaction with usaspending.gov
}
\examples{
\dontrun{
usaspending_api_call(unique_id="XXXXX_XXXXXX", type="funding", verbose =T)}

}
