% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{validate_savings}
\alias{validate_savings}
\title{validate_savings}
\usage{
validate_savings(merged_data = NULL, outlay_data = NULL)
}
\arguments{
\item{merged_data}{Data created by clean_and_match()}

\item{outlay_data}{Include data created by get_usaspending_outlay()}
}
\value{
A data.frame
}
\description{
A utility function to take merged data (clean_and_match()) and run
mathematical and comparative functions to enable cross-referencing with
DOGE-supplied data.
}
\examples{
\dontrun{
validate_savings(M,O)
}

}
