% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{xml_to_table}
\alias{xml_to_table}
\title{xml_to_table}
\usage{
xml_to_table(
  x = NULL,
  xtype = NULL,
  include_related = FALSE,
  primary_PIID = NULL
)
}
\arguments{
\item{x}{XML data}

\item{xtype}{Type value generated by check_xml_type()}

\item{include_related}{Logical, whether to include related values.}

\item{primary_PIID}{String, the PIID value of the parent, or primary award lookup.}
}
\value{
A dataframe
}
\description{
Utility function to convert XML data to dataframe; utilized by fpds_get_data().
}
\examples{
\dontrun{
xml_to_table(xml_data, "IDV", include_related=FALSE)
}

}
