% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{clean_and_match}
\alias{clean_and_match}
\title{clean_and_match}
\usage{
clean_and_match(
  FPDS_data = NULL,
  DOGE_data = NULL,
  outlay_data = NULL,
  print_status = TRUE,
  verbose = F
)
}
\arguments{
\item{FPDS_data}{Supplied data from FPDS}

\item{DOGE_data}{Supplied data from DOGE}

\item{outlay_data}{A dataset from get_usaspending_outlay(), to add to merge file.}

\item{print_status}{Logical, whether to return feedback}

\item{verbose}{Logical, whether to return messages}
}
\value{
A merged dataframe.
}
\description{
A utility function designed to help match DOGE data and FPDS data. FPDS data
returns multiple transaction values per PIID; DOGE data only includes on row.
While it is possible to match on PIID and modNumber, this frequently returns
several values.
}
\examples{
\dontrun{
clean_and_match(F,D,print_status=TRUE)
}
}
