% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{create_vendor_lookup}
\alias{create_vendor_lookup}
\title{create_vendor_lookup}
\usage{
create_vendor_lookup(DOGE = NULL, FPDS = NULL)
}
\arguments{
\item{DOGE}{supplied DOGE data}

\item{FPDS}{supplied FPDS data}
}
\value{
a data.frame
}
\description{
Internal function to create standardized vendor lookup table.
}
\examples{
\dontrun{
create_vendor_lookup(d,f)}

}
