% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doge.R
\name{doge_get_data}
\alias{doge_get_data}
\title{DOGE Get Data}
\usage{
doge_get_data(verbose = TRUE, saveFile = FALSE, saveDir = NULL)
}
\arguments{
\item{verbose}{Logical, whether to produce messages}

\item{saveFile}{Logical, whether to save file}

\item{saveDir}{File dirctory path for saving files.}
}
\value{
A named list comprised of dataframes.
}
\description{
A fairly straightforward scraper that pulls data from dynamically-created
scripts (javascript), cleans, and processes in table form.
}
\examples{
\dontrun{
doge_get_data()
}
}
