% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{doge_get_data}
\alias{doge_get_data}
\title{doge_get_data}
\usage{
doge_get_data(
  selectGrants = TRUE,
  selectContracts = TRUE,
  selectLeases = TRUE,
  selectPayments = TRUE,
  limit = 500,
  pageNumStart = 1,
  verbose = T
)
}
\arguments{
\item{selectGrants}{Logical, whether to select grants.}

\item{selectContracts}{Logical, whether to select contracts}

\item{selectLeases}{Logical, whether to select leases.}

\item{selectPayments}{Logical, whether to select payments.}

\item{limit}{Defaults to 500; how many entries per page.}

\item{pageNumStart}{Which page to start on. Defaults to 1.}

\item{verbose}{Logical, whether to provide feedback.}
}
\value{
A data.frame
}
\description{
A simple function to interact with api.doge.gov. This was built for api version 0.0.1.-beta.
}
\examples{
\dontrun{
doge_get_data()
}
}
