% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpds.R
\name{fpds_get_data}
\alias{fpds_get_data}
\title{Get FPDS Data}
\usage{
fpds_get_data(
  piid = NULL,
  return = "xml",
  include_related = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{piid}{A PIID value or vector of values.}

\item{return}{If "raw", output raw HTML file.}

\item{verbose}{Logical, to display progressbar.}
}
\value{
A named list, consisting of two dataframes: AWARDS and IDV.
}
\description{
An iterative function to query FPDS data by PIID using FPDS Atom Feed. See
https://www.fpds.gov/fpdsng_cms/index.php/en/worksite for more information.
}
\examples{
\dontrun{
fpds_get_data(piid="XXXXXXXXX")
}
}
