% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{generate_unique_key}
\alias{generate_unique_key}
\title{generate_unique_key}
\usage{
generate_unique_key(
  PIID = NULL,
  toptier_code = NULL,
  award_type = NULL,
  idvPIID = NULL,
  idvAgency = NULL
)
}
\arguments{
\item{PIID}{Character string, the PIID value of an award.}

\item{toptier_code}{An agency code}

\item{award_type}{Either "contract","grant","IDV"}

\item{idvPIID}{Character string, the IDV PIID value of an award.}

\item{idvAgency}{Character string, the agency number of an IDV.}
}
\value{
A character string
}
\description{
A utility function to generate a unique key as used at usaspending.gov.
}
\examples{
\dontrun{
generate_unique_key(PIID="XXX999333",toptier_code="0900",award_type="grant")
}
}
